feat: allow logging of all item changes to separate logger
Push all item changes to a separate logger globally configurable in the
application config. We put configuration like logger name, level and
template on item level for maximum flexibility, so we could exclude
specific items or override the log level per item. I'm pretty sure that
this level of flexibility is not needed, in which case some of this code
could get migrated to class Items later on.

Closes #2